Refactor to use class for link
